+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+
+## 8.17.2
+- Added possibility to sort a player's inventory from console using /invsort <player> [toggle|on|off|hotbar|inv|all]
+- Added placeholders (see new config.yml)
+- Added "use permissions" metric
+
 ## 8.17.1
 - Improved sorting of colored blocks
   - White, light gray, gray and black comes first